Started moving NLP code from services.py to services_nlp.py
